migrated everything to their header and related files (build is unstable)
