+++ kindaRPGthing.vcxproj.filters
@@ -18,9 +18,18 @@
     <ClCompile Include="GameFile.c">
       <Filter>來源檔案</Filter>
     </ClCompile>
+    <ClCompile Include="GameItems.c">
+      <Filter>來源檔案</Filter>
+    </ClCompile>
+    <ClCompile Include="Mob_UsrStruct.c">
+      <Filter>來源檔案</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="itemActions.h">
+      <Filter>標頭檔</Filter>
+    </ClInclude>
+    <ClInclude Include="BattleMechanics.h">
       <Filter>標頭檔</Filter>
     </ClInclude>
   </ItemGroup>
